requirements.txt
streamlit
sentence-transformers
faiss-cpu
langchain
PyPDF2
numpy
llama-cpp-python
transformers
